Generated a template for dockerized Lambda function (#13)
* feat: added Dockerized Lambda function for test

With this commit the resources and the infrastructure for deploying a test application on AWS Lambda function with Docker image are added.
This could be a starting point for other Lambda applications so this code can be deleted in future if no more needed.

* feat: created GH Action for dispatched trigger

* feat: added new IAM role for ECR access

With this commit, a new IAM role with the related policies is added. This new role is used to permits the full access to ECR by pipelines.

* fix: updated repository_dispatch action

* feat: added ECR repository resource

With this commit a new ECR repository resource is added. Also, the resource container_registry_login used for logging in AWS ECR is removed due to the fact that is no more needed.